splitting files refs #33430
