+++ Dockerfile
@@ -13,3 +13,5 @@ WORKDIR /src
 COPY requirements-pip.txt /src
 
 RUN pip3 install -r requirements-pip.txt
+
+ENV PYTHONPATH "${PYTHONPATH}/src"
