+++ assets
@@ -1,1 +1,1 @@
-./submodules/mediasite_client/assets
+submodules/mediasite_client/assets/
